Added subfolder listing capacity to the Folder Item.

+++ fake-starteam/src/com/starbase/starteam/Folder.java
@@ -19,7 +19,10 @@ package com.starbase.starteam;
 import java.io.File;
 import java.io.FileInputStream;
 import java.io.FileOutputStream;
+import java.io.FilenameFilter;
 import java.io.IOException;
+import java.util.ArrayList;
+import java.util.List;
 import java.util.Properties;
 
 import org.ossnoize.fakestarteam.InternalPropertiesProvider;
@@ -29,9 +32,26 @@ import org.ossnoize.fakestarteam.exception.InvalidOperationException;
 public class Folder extends Item {
 	
 	private static final String FOLDER_PROPERTIES = "folder.properties";
+	private static final FilenameFilter FOLDER_TESTER = new FilenameFilter() {
+		@Override
+		public boolean accept(File dir, String name) {
+			return name.equalsIgnoreCase(FOLDER_PROPERTIES);
+		}
+	};
 
 	public Folder(Server server) {
 		throw new UnsupportedOperationException("Unknown goal for this constructor");
+	}
+	
+	public Folder(Folder parent, String name, String workingFolder) {
+		try {
+			String folder = parent.holdingPlace.getCanonicalPath() + File.separator + name;
+			holdingPlace = new File(folder);
+			validateHoldingPlace();
+			loadFolderProperties();
+		} catch (IOException e) {
+			e.printStackTrace();
+		}
 	}
 	
 	protected Folder(View currentView) {
@@ -43,22 +63,29 @@ public class Folder extends Item {
 			String rootFolder = serverArchive.getCanonicalPath() + File.separator +
 					currentView.getProject().getName() + File.separator + currentView.getName();
 			holdingPlace = new File(rootFolder);
-			if(holdingPlace.exists()) {
-				if(holdingPlace.isFile()) {
-					holdingPlace.delete();
-					holdingPlace.mkdirs();
-				}
-			} else {
-				holdingPlace.mkdirs();
-			}
+			validateHoldingPlace();
+			loadFolderProperties();
 		} catch (IOException e) {
 			e.printStackTrace();
 		}
-		loadFolderProperties();
 		// We don't want the root folder to have any default name.
 		// so overwrite the view name with a blank one and update.
 		setName("");
 		update();
+	}
+
+	private void validateHoldingPlace() {
+		if(null == holdingPlace) {
+			throw new InvalidOperationException("Cannot create a folder without an holding place.");
+		}
+		if(holdingPlace.exists()) {
+			if(holdingPlace.isFile()) {
+				holdingPlace.delete();
+				holdingPlace.mkdirs();
+			}
+		} else {
+			holdingPlace.mkdirs();
+		}
 	}
 	
 	public void setName(java.lang.String name) {
@@ -73,6 +100,20 @@ public class Folder extends Item {
 			throw new InvalidOperationException("The properties are not initialized");
 		}
 		return itemProperties.getProperty(propertyKeys.FOLDER_NAME);
+	}
+	
+	public Folder[] getSubFolders() {
+		List<Folder> generatedList = new ArrayList<Folder>();
+		for(File f : holdingPlace.listFiles()) {
+			if(f.isDirectory()) {
+				String[] subFolders = f.list(FOLDER_TESTER);
+				if(subFolders != null && subFolders.length == 1) {
+					generatedList.add(new Folder(this, f.getName(), ""));
+				}
+			}
+		}
+		Folder[] buffer = new Folder[generatedList.size()];
+		return generatedList.toArray(buffer);
 	}
 	
 	@Override
